feat: add build_feature function to __init__.py for user and diner feature engineering
